Renamed 'Analyze' app to 'api' to follow best practices for naming and to clarify that this contains our endpoints for handling requests
