+++ api/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 from rest_framework.urlpatterns import format_suffix_patterns
-from Analyzer import views
+from api import views
 
 urlpatterns = [
     path('stocks/', views.StockList.as_view()),
